optimal solution for largest_container_area

+++ src/arrays/most_water_container/WaterContainer.java
@@ -1,6 +1,6 @@
 package arrays.most_water_container;
 
 public interface WaterContainer {
-    double getLargestContainerAreaV1(int[] arr);
-    double getLargestContainerAreaV2(int[] arr);
+    int getLargestContainerAreaV1(int[] arr);
+    int getLargestContainerAreaV2(int[] arr);
 }
